message unify nonce interface and return encoded msg key

+++ contracts/bridge/src/interfaces/IOutboundLane.sol
@@ -26,5 +26,6 @@ interface IOutboundLane {
     /// At the beginning of the launch, submmiter is permission, after the system is stable it will be permissionless.
     /// @param targetContract The target contract address which you would send cross chain message to
     /// @param encoded The calldata which encoded by ABI Encoding `abi.encodePacked(SELECTOR, PARAMS)`
-    function send_message(address targetContract, bytes calldata encoded) external payable returns (uint64);
+    /// @return Encoded message key
+    function send_message(address targetContract, bytes calldata encoded) external payable returns (uint256);
 }
